refactor: gained ability to be used as subproject
e.g. via add_subdirectory or FetchContent
