+++ fuzzing/cjson_read_fuzzer.c
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-#include "../cJSON.h"
+#include "../cjson/cJSON.h"
 
 int LLVMFuzzerTestOneInput(const uint8_t* data, size_t size); /* required by C89 */
